Repaired a bug when expanding the macros for isaac_rand
Seems like the expansion for (a? b: c,d) is something like
(a? b:c) and in both cases do d.
So we just jumped the index of wheel2 over the odd numbers,
never reaching a 0, and thus never recalculating the new
values.

This surely will mean that this is not the fastest generator,
as it was with this bug.