Added the 64 bit version of ISAAC generator
It has been easier to just 'guess' the needed
changes than trying to understand the 'code'
the author has in his webpage.

At least, the output for both of them are the
same, so we should think that this code is
correct.

As an interesting fact, this version produces
numbers faster than parisi 32 bits, which
was the fastest until now.