Edited the changelog with some information about today's experiments.
The experiment has been to change the pr_rand() function
to a pr_calc() function and a define that wraps a
reference to the wheel of random numbers from
pr_default_state if the index is not 0 or a call to pr_calc
and then a reference to the whell. Just as it is done
for ISAAC.

The results show that almost none improvement in speed
are achieved, so the code is left as is. Also, changing
the function to a define will surely break something.
So it's best left as it is now (^_^)