Initial commit for the 'distributions' branch
In this branch I will try to add support for generating
random numbers according to different probability
distributions.

There's still work to do about 'how the 'main' library is
going to make use of this?'. This is yet to be decided.
What I'm thinking is a MACRO in rands_config.h that allows
you to choose whether you want to use the distributions
or not. Of course, I still have to figure out how to make
all that work...

Fo the initial release, we have three distributions:
 - Normal: box-muller and polard methods
 - Poisson
 - Chi²