Added the binomial distribution
This implements the naïve algorithm for the binomial distribution:
Just throw 'n' times a random number and return the number of
times that the number is less than p.

There are two versions of the algorithm. The first one would be
the "usual" one, with p a number between 0 and 1. The second
one receives a probability p as an unsigned between 0 and UINT_MAX.
This can speed up comparations in the algorithm as it would need
'n' less divisions by UINT_MAX. Usually, the optimizer of the
compiler will make the two run equally well, but if there isn't
one present, it may speed up the generation.