Added a test for the distributions library
The test doesn't check that the numbers are correct,
it just calculates them and shows them along with
the theoretical ones. Then the user must compare
them and accept the results or not.

This is the test that was used to discover the bug
from last commit (^_^)

This commit is just about the tests, there's still
work to do to unify both libraries and to have a
'Preferred Generator'