Added full code for SSE2 version of parisi
To be able of make more than one calculations at a time,
I use the trick of Isaac generator: keep two wheels, one
with the 'true' state of the generator, and another with
the future output. Thus, I can calculate in advance 256
random numbers making use of SSE and after that returning
them as they are needed.

Although this may seem simple, care must be taken with
the wheel's bounds and the positions from where we have
to combine numbers, as they are not multiple of 4.

This can be seen when I have to calculate 12 random numbers
one at a time instead of using SSE. This is because the
first three (52-54) can't be done with SSE, we could do
the next four (55-58) in one SSE step, but the next five
(59-63) have to be done one at a time.