Added support for SSE2 in Mersenne generator
In this case, I have not divided Mersenne in two different
generators, one with SSE2 and the other without. This is
because, in this case, the original algorithm already make
use of a second 'wheel' of numbers that gets filled with
the numbers that will be outputed.

So in this case, instead of adding it and creating the
mt_calc functions, I have simply rewritten it with SSE.