Added support for SSE4.1 in Mersenne Twister
With the SSE4.1 blend function we can avoid all the 'magic'
we had to do to obtain the correct 'magic number' for
the xors.

Strangely, using this version speeds up (for me) the speed
test for 10M from ~725 ms to ~700 ms, but for 20M is slower,
from ~1400 ms to ~1450 ms.

The original behaviour can be obtained by adding '-mno-sse4.1'
to the variable CFLAGS in the Makefile